Restructure errors.py to move to akd/_base/errors.py
akd.errors is jsut a public proxy
